# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AnomalyStatus": {
        "enum": [
          "open",
          "reviewed",
          "cleared"
        ],
        "title": "AnomalyStatus",
        "type": "string"
      },
...
      "ForecastResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ForecastItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ForecastResponse",
        "type": "object"
      },
...
      "LifecycleStatus": {
        "enum": [
          "open",
          "acknowledged",
          "cleared"
        ],
        "title": "LifecycleStatus",
        "type": "string"
      },
...
      "Page_AlarmItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlarmItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total",
          "items"
        ],
        "title": "Page[AlarmItem]",
        "type": "object"
      },
...
      "Severity": {
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Severity",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "energy_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Monitoring and advisory only; no industrial load-control path.",
    "title": "Intelligent Industrial Energy Dashboard API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/alarms": {
      "get": {
        "operationId": "alarms_api_v1_alarms_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LifecycleStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Severity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "device_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AlarmItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_DATE_RANGE",
                    "details": {
                      "field": "to"
                    },
                    "message": "The end timestamp must be after the start timestamp.",
                    "request_id": "01J5EXAMPLE"
                  }
                }
              }
            },
            "description": "Invalid query parameters"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Alarms",
        "tags": [
          "dashboard reads"
        ]
      }
    },
...
    "/api/v1/anomalies": {
      "get": {
        "operationId": "anomalies_api_v1_anomalies_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AnomalyStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Severity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "device_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AnomalyItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_DATE_RANGE",
                    "details": {
                      "field": "to"
                    },
                    "message": "The end timestamp must be after the start timestamp.",
                    "request_id": "01J5EXAMPLE"
                  }
                }
              }
            },
            "description": "Invalid query parameters"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Anomalies",
        "tags": [
          "dashboard reads"
        ]
      }
    },
...
    "/api/v1/dashboard/summary": {
      "get": {
        "operationId": "dashboard_summary_api_v1_dashboard_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "timezone",
            "required": false,
            "schema": {
              "default": "Africa/Johannesburg",
              "title": "Timezone",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_DATE_RANGE",
                    "details": {
                      "field": "to"
                    },
                    "message": "The end timestamp must be after the start timestamp.",
                    "request_id": "01J5EXAMPLE"
                  }
                }
              }
            },
            "description": "Invalid query parameters"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Dashboard Summary",
        "tags": [
          "dashboard reads"
        ]
      }
    },
...
    "/api/v1/forecasts/latest": {
      "get": {
        "operationId": "latest_forecast_api_v1_forecasts_latest_get",
        "parameters": [
          {
            "in": "query",
            "name": "horizon_minutes",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Horizon Minutes",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "device_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_DATE_RANGE",
                    "details": {
                      "field": "to"
                    },
                    "message": "The end timestamp must be after the start timestamp.",
                    "request_id": "01J5EXAMPLE"
                  }
                }
              }
            },
            "description": "Invalid query parameters"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Latest Forecast",
        "tags": [
          "dashboard reads"
        ]
      }
    },
...
  }
}